refactor(render): Move preprocess logic out
