+++ src/renderer/render.rs
@@ -13,16 +13,16 @@ use anstyle::Style;
 use super::DecorStyle;
 use super::Renderer;
 use super::margin::Margin;
+use super::preprocess::{DisplaySuggestion, Preprocessed, PreprocessedElement, PreprocessedGroup};
 use super::stylesheet::Stylesheet;
 use crate::level::{Level, LevelInner};
 use crate::renderer::source_map::{
-    AnnotatedLineInfo, LineInfo, Loc, SourceMap, SplicedLines, SubstitutionHighlight, TrimmedPatch,
+    AnnotatedLineInfo, LineInfo, Loc, SourceMap, SplicedLines, SubstitutionHighlight,
 };
 use crate::renderer::styled_buffer::StyledBuffer;
 use crate::snippet::Id;
 use crate::{
-    Annotation, AnnotationKind, Element, Group, Message, Origin, Padding, Patch, Report, Snippet,
-    Title,
+    Annotation, AnnotationKind, Element, Group, Message, Origin, Patch, Report, Snippet, Title,
 };
 
 const ANONYMIZED_LINE_NUM: &str = "LL";
@@ -2499,39 +2499,6 @@ impl LineAnnotation<'_> {
     }
 }
 
-#[derive(Clone, Copy, Debug)]
-pub(crate) enum DisplaySuggestion {
-    Underline,
-    Diff,
-    None,
-    Add,
-}
-
-impl DisplaySuggestion {
-    fn new(complete: &str, patches: &[TrimmedPatch<'_>], sm: &SourceMap<'_>) -> Self {
-        let has_deletion = patches
-            .iter()
-            .any(|p| p.is_deletion(sm) || p.is_destructive_replacement(sm));
-        let is_multiline = complete.lines().count() > 1;
-        if has_deletion && !is_multiline {
-            Self::Diff
-        } else if patches.len() == 1
-            && patches.first().is_some_and(|p| {
-                p.replacement.ends_with('\n') && p.replacement.trim() == complete.trim()
-            })
-        {
-            // We are adding a line(s) of code before code that was already there.
-            Self::Add
-        } else if (patches.len() != 1 || patches[0].replacement.trim() != complete.trim())
-            && !is_multiline
-        {
-            Self::Underline
-        } else {
-            Self::None
-        }
-    }
-}
-
 // We replace some characters so the CLI output is always consistent and underlines aligned.
 // Keep the following list in sync with `rustc_span::char_width`.
 const OUTPUT_REPLACEMENTS: &[(char, &str)] = &[
@@ -2663,185 +2630,6 @@ enum TitleStyle {
     Secondary,
 }
 
-struct Preprocessed<'a> {
-    max_line_num: Option<usize>,
-    report_primary_path: Option<&'a Cow<'a, str>>,
-    groups: Vec<PreprocessedGroup<'a>>,
-}
-
-impl Preprocessed<'_> {
-    fn preprocess<'a>(groups: &'a [Group<'a>]) -> Preprocessed<'a> {
-        preprocess(groups)
-    }
-}
-
-struct PreprocessedGroup<'a> {
-    group: &'a Group<'a>,
-    elements: Vec<PreprocessedElement<'a>>,
-    primary_path: Option<&'a Cow<'a, str>>,
-    max_depth: usize,
-}
-
-enum PreprocessedElement<'a> {
-    Message(&'a Message<'a>),
-    Cause(
-        (
-            &'a Snippet<'a, Annotation<'a>>,
-            SourceMap<'a>,
-            Vec<AnnotatedLineInfo<'a>>,
-        ),
-    ),
-    Suggestion(
-        (
-            &'a Snippet<'a, Patch<'a>>,
-            SourceMap<'a>,
-            SplicedLines<'a>,
-            DisplaySuggestion,
-        ),
-    ),
-    Origin(&'a Origin<'a>),
-    Padding(Padding),
-}
-
-fn preprocess<'a>(groups: &'a [Group<'a>]) -> Preprocessed<'a> {
-    let mut max_line_num = None;
-    let mut report_primary_path = None;
-    let mut out = Vec::with_capacity(groups.len());
-    for group in groups {
-        let mut elements = Vec::with_capacity(group.elements.len());
-        let mut primary_path = None;
-        let mut max_depth = 0;
-        for element in &group.elements {
-            match element {
-                Element::Message(message) => {
-                    elements.push(PreprocessedElement::Message(message));
-                }
-                Element::Cause(cause) => {
-                    let sm = SourceMap::new(&cause.source, cause.line_start);
-                    let (depth, annotated_lines) =
-                        sm.annotated_lines(cause.markers.clone(), cause.fold);
-
-                    if cause.fold {
-                        let end = cause
-                            .markers
-                            .iter()
-                            .map(|a| a.span.end)
-                            .max()
-                            .unwrap_or(cause.source.len())
-                            .min(cause.source.len());
-
-                        if cause.line_numbering {
-                            max_line_num = Some(max(
-                                cause.line_start + newline_count(&cause.source[..end]),
-                                max_line_num.unwrap_or(0),
-                            ));
-                        }
-                    } else {
-                        if cause.line_numbering {
-                            max_line_num = Some(max(
-                                cause.line_start + newline_count(&cause.source),
-                                max_line_num.unwrap_or(0),
-                            ));
-                        }
-                    }
-
-                    if primary_path.is_none() {
-                        primary_path = Some(cause.path.as_ref());
-                    }
-                    max_depth = max(depth, max_depth);
-                    elements.push(PreprocessedElement::Cause((cause, sm, annotated_lines)));
-                }
-                Element::Suggestion(suggestion) => {
-                    let sm = SourceMap::new(&suggestion.source, suggestion.line_start);
-                    if let Some(spliced_lines) =
-                        sm.splice_lines(suggestion.markers.clone(), suggestion.fold)
-                    {
-                        let display_suggestion = DisplaySuggestion::new(
-                            &spliced_lines.complete,
-                            &spliced_lines.patches,
-                            &sm,
-                        );
-
-                        if suggestion.line_numbering {
-                            if suggestion.fold {
-                                if let Some(first) = spliced_lines.patches.first() {
-                                    let (l_start, _) =
-                                        sm.span_to_locations(first.original_span.clone());
-                                    let nc = newline_count(&spliced_lines.complete);
-                                    let sugg_max_line_num = match display_suggestion {
-                                        DisplaySuggestion::Underline => l_start.line,
-                                        DisplaySuggestion::Diff => {
-                                            let file_lines = sm.span_to_lines(first.span.clone());
-                                            file_lines
-                                                .last()
-                                                .map_or(l_start.line + nc, |line| line.line_index)
-                                        }
-                                        DisplaySuggestion::None => l_start.line + nc,
-                                        DisplaySuggestion::Add => l_start.line + nc,
-                                    };
-                                    max_line_num =
-                                        Some(max(sugg_max_line_num, max_line_num.unwrap_or(0)));
-                                }
-                            } else {
-                                max_line_num = Some(max(
-                                    suggestion.line_start + newline_count(&spliced_lines.complete),
-                                    max_line_num.unwrap_or(0),
-                                ));
-                            }
-                        }
-
-                        elements.push(PreprocessedElement::Suggestion((
-                            suggestion,
-                            sm,
-                            spliced_lines,
-                            display_suggestion,
-                        )));
-                    }
-                }
-                Element::Origin(origin) => {
-                    if primary_path.is_none() {
-                        primary_path = Some(Some(&origin.path));
-                    }
-                    elements.push(PreprocessedElement::Origin(origin));
-                }
-                Element::Padding(padding) => {
-                    elements.push(PreprocessedElement::Padding(padding.clone()));
-                }
-            }
-        }
-        let group = PreprocessedGroup {
-            group,
-            elements,
-            primary_path: primary_path.unwrap_or_default(),
-            max_depth,
-        };
-        if report_primary_path.is_none() && group.primary_path.is_some() {
-            report_primary_path = group.primary_path;
-        }
-        out.push(group);
-    }
-
-    Preprocessed {
-        max_line_num,
-        report_primary_path,
-        groups: out,
-    }
-}
-
-fn newline_count(body: &str) -> usize {
-    #[cfg(feature = "simd")]
-    {
-        // Trailing newlines do not count towards the number of lines
-        // (this is based into `str::lines`)
-        let trailing_newline = body.ends_with('\n');
-        memchr::memchr_iter(b'\n', body.as_bytes()).count() - usize::from(trailing_newline)
-    }
-    #[cfg(not(feature = "simd"))]
-    {
-        body.lines().count().saturating_sub(1)
-    }
-}
-
 struct Hyperlink<D: fmt::Display> {
     url: Option<D>,
 }
@@ -2873,7 +2661,7 @@ impl<D: fmt::Display> fmt::Display for Hyperlink<D> {
 
 #[cfg(test)]
 mod test {
-    use super::{OUTPUT_REPLACEMENTS, newline_count};
+    use super::OUTPUT_REPLACEMENTS;
     use snapbox::IntoData;
 
     fn format_replacements(replacements: Vec<(char, &str)>) -> String {
@@ -2894,35 +2682,5 @@ mod test {
         let expected = format_replacements(expected);
         let actual = format_replacements(OUTPUT_REPLACEMENTS.to_owned());
         snapbox::assert_data_eq!(actual, expected.into_data().raw());
-    }
-
-    #[test]
-    fn ensure_newline_count_correct() {
-        let source = r#"
-                cargo-features = ["path-bases"]
-
-                [package]
-                name = "foo"
-                version = "0.5.0"
-                authors = ["wycats@example.com"]
-
-                [dependencies]
-                bar = { base = '^^not-valid^^', path = 'bar' }
-            "#;
-        assert_eq!(newline_count(source), 10);
-
-        assert_eq!(newline_count(""), 0);
-
-        assert_eq!(newline_count("one"), 0);
-
-        assert_eq!(newline_count("one\n"), 0);
-
-        assert_eq!(newline_count("one\ntwo"), 1);
-
-        assert_eq!(newline_count("one\ntwo\n"), 1);
-
-        assert_eq!(newline_count("one\n\n"), 1);
-
-        assert_eq!(newline_count("one\r\ntwo\r\n"), 1);
     }
 }
